% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayconpar.R
\name{bayconResultClass}
\alias{bayconResultClass}
\title{Multi-result class for baycon}
\usage{
bayconResultClass(pEstimatesList = NULL, pEstimatesListMedian = NULL,
  pEstimatesListMode = NULL, var = NULL, mvnErrorDistList = NULL)
}
\arguments{
\item{pEstimatesList}{mean estimates from Stan}

\item{var}{variance estimates from Stan}

\item{mvnErrorDistList}{MVN approx to posterior}
}
\description{
Create class which holds multiple results for each loop iteration.
Each loop iteration populates three properties.
For a great tutorial on \code{S3} classes, see:
http://www.cyclismo.org/tutorial/R/s3Classes.html#creating-an-s3-class
}
