% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-across.R
\name{compute_across_model_decon}
\alias{compute_across_model_decon}
\title{Compute deconvolution from across-cell type modeling based on the
cell type ratios derived from \code{compute_split_proportions}}
\usage{
compute_across_model_decon(x, cluster_labels, convolved, verbose = TRUE)
}
\arguments{
\item{x}{A \code{GRanges} object returned from the function \code{compute_split_proportions}}

\item{cluster_labels}{A \code{data.frame} or \code{GRanges} object containing
genomic bins associated with a given cluster label}

\item{convolved}{a \code{GRanges} object with convolved ChIP-seq signal data}

\item{verbose}{boolean indicating whether to flush output to console}
}
\value{
a \code{GRanges} object containing metadata of across-cell type
deconvolution
}
\description{
Compute deconvolution from across-cell type modeling based on the
cell type ratios derived from \code{compute_split_proportions}
}
