% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-across.R
\name{compute_split_proportions}
\alias{compute_split_proportions}
\title{Compute cell type ratios to split convolved signal based on across-cell type
predictions}
\usage{
compute_split_proportions(x)
}
\arguments{
\item{x}{a list containing cluster-specific, average loci intensity predictions,
where each element in the list corresponds to a cluster.}
}
\value{
a list containing predictions that are convered to split proportions
}
\description{
Compute cell type ratios to split convolved signal based on across-cell type
predictions
}
