% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconR-package.R
\docType{package}
\name{deconR-package}
\alias{deconR-package}
\alias{deconR}
\title{The 'deconR' package.}
\description{
PACKAGE for bayesian deconvolution of bulk RNAseq data
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
