% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dgt}
\alias{dgt}
\title{Density of a generalized t-distribution}
\usage{
dgt(x, df = 1, mu = 0, sigma = 1)
}
\arguments{
\item{x}{sample value}

\item{mu}{the location parameter}

\item{sigma}{the scale parameter}
}
\value{
the density of the t-distribution at x
}
\description{
This produces the density for a generalized t-distribution.
Details of the distribution are here -
https://en.wikipedia.org/wiki/Student\%27s_t-distribution#Generalized_Student's_t-distribution .
We want this density to match the one used in rstan here -
https://mc-stan.org/docs/2_18/functions-reference/student-t-distribution.html#probability-density-function-3
}
