% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode-data.R
\docType{data}
\name{encode_expr}
\alias{encode_expr}
\title{Mouse ENCODE RNA-seq gene expression values}
\format{An object of class \code{"data.frame"}.}
\source{
\href{http://www.mouseencode.org}{ENCODE Mouse Project}
}
\usage{
data(encode_expr)
}
\description{
Data from 87 mouse cell types source from ENCODE contains RNA-seq depth-normalized
TPM values with replicates aggregrated for 14539 genes.
}
\examples{
data(encode_expr)
}
\references{
Yue et al. (2014) Nature 515:355-364
(\href{https://www.ncbi.nlm.nih.gov/pubmed/25409824}{PubMed})
}
\keyword{datasets}
