% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode-data.R
\docType{data}
\name{encode_signal}
\alias{encode_signal}
\title{Mouse ENCODE H3K27ac raw signal values}
\format{An object of class \code{"GRanges"}; see \code{\link[GenomicRanges]{makeGRangesFromDataFrame}}.}
\source{
\href{http://www.mouseencode.org}{ENCODE Mouse Project}
}
\usage{
data(encode_signal)
}
\description{
Data from 87 mouse cell types source from ENCODE contains scaled (log2) H3K27ac
signal indensity values with a genomic resolution of 200 bps covering all
of chr16 (491038 non-overlapping 200bp bins).
}
\examples{
data(encode_signal)
}
\references{
Yue et al. (2014) Nature 515:355-364
(\href{https://www.ncbi.nlm.nih.gov/pubmed/25409824}{PubMed})
}
\keyword{datasets}
