% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_mode}
\alias{estimate_mode}
\title{Estimate the mode of the univarite distribution}
\usage{
estimate_mode(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
the mode of x
}
\description{
Surprisingly, R does not have a mode function,
well it does but it doesn't do what you think it does.
This implmentation was suggested in the SO answer here:
https://stackoverflow.com/a/13874750/2868715
}
