% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_convolved}
\alias{make_convolved}
\title{Generate an in silico convolved ChIP-seq signal track from known ChIP-seq data}
\usage{
make_convolved(x, log2_scale = TRUE)
}
\arguments{
\item{x}{a list of \code{GRanges} objects with ChIP-seq signal data from different cell types to
to be convolved}

\item{log2_scale}{boolean indicating whether to log2 scale (with pseudocount of 1)}
}
\value{
a \code{GRanges} object with convolved ChIP-seq signal data
}
\description{
Generate an in silico convolved ChIP-seq signal track from known ChIP-seq data
}
