% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSimDataset.R
\name{makeSimDataset}
\alias{makeSimDataset}
\title{Function to make a simple simulated dataset for differential expression testing.}
\usage{
makeSimDataset(sdCancer = 1, sdNormal = 1, seed = 12345,
  theProp = p_theProp, propInv = p_propInv)
}
\arguments{
\item{sdCancer}{std dev of cancer sample}

\item{sdNormal}{std dev of normal sample}

\item{seed}{set the seed for the simulation}
}
\value{
An object of type list with following names:

\code{cancerExpressionSimMat, normalExpressionSimMat, bulkExpressionSimMat,
simulatedEffectSizesCancer, simulatedEffectSizesNormal}
}
\description{
The dataset created has 600 genes x 1000 samples and needs the true proportion values.
This works for two cell types - cancer and normal.
NOT to be exported.
}
