% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-across.R
\name{model_across_celltypes}
\alias{model_across_celltypes}
\title{Predicts ChIP-seq profiles from gene expression by learned relationships between RNA and
ChIP within ENCODE data}
\usage{
model_across_celltypes(clusters, scrna_aggr, genes = NULL, encode_signal,
  encode_expr, data_split, output_dir = NULL, seed = 42)
}
\arguments{
\item{clusters}{A \code{data.frame} containing genomic bins associated with a given
cluster label}

\item{scrna_aggr}{\code{data.frame} containing aggregrated single cell gene counts}

\item{genes}{a character vector of genes to be used as features to the model}

\item{encode_signal}{An object of class \code{"GRanges"} containing normalized
ChIP-seq signal indensity values sourced from Mouse ENCODE}

\item{encode_expr}{An object of class \code{"data.frame"} containing normalized
RNA-seq gene expression information source from Mouse ENCODE}

\item{data_split}{a list of cell types to be used for TRAINING and VALIDATION}

\item{output_dir}{optional; if specified, then function will output model/loci-specific
scatterplots with modeling statistics to the specified directory}

\item{seed}{set seed for model reproducibility}
}
\value{
a list of lists contaning predicted and true, cluster-averaged predictions for each cell
type. Also included are average prediction from using scRNA aggregrates.
}
\description{
Predicts ChIP-seq profiles from gene expression by learned relationships between RNA and
ChIP within ENCODE data
}
