% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_model_results}
\alias{plot_model_results}
\title{Plot a scatterplot of model predicted values against true values}
\usage{
plot_model_results(x, bins, title = "")
}
\arguments{
\item{x}{a \code{data.frame} contains model predictions}

\item{bins}{a \code{GRanges} object containing the genomic bins within a
given cluster of loci. This object is used for cluster and n_loci lookup.}

\item{title}{plot title/identifier}
}
\value{
a \code{ggplot} scatterplot object
}
\description{
Plot a scatterplot of model predicted values against true values
}
