% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_signal}
\alias{read_signal}
\title{Read in ChIP-seq signal intensity values}
\usage{
read_signal(x, format = "bw", genome = "mm10", chr = NULL)
}
\arguments{
\item{x}{file name of ChIP-seq data}

\item{format}{file format of ChIP-seq data}

\item{genome}{If not missing, should be “wig” or “bw” (case insensitive).}

\item{chr}{optional; specify to subset to given chromosome}
}
\value{
a \code{GRanges} object with ChIP-seq signal data
}
\description{
Read in ChIP-seq signal intensity values
}
