% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rgt}
\alias{rgt}
\title{Samples from a generalized t-distribution}
\usage{
rgt(n, df = 1, mu = 0, sigma = 1)
}
\arguments{
\item{n}{number of samples to be generated}

\item{df}{degrees of freedom}

\item{mu}{the location parameter}

\item{sigma}{the scale parameter}
}
\value{
a vector of generalized student-t distributed random variables
}
\description{
Applies the specified location and scale paramters to the standart t-distribution.
The mean of this distribution is mu,
the variance is sigma^2*(df/(df - 2))
}
