% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample.simplex}
\alias{sample.simplex}
\title{Sample from a simplex}
\usage{
sample.simplex(n = 4, safe = T)
}
\arguments{
\item{n}{number of dimensions}

\item{safe}{\code{logical}, if \code{TRUE} ensures there are no zeroes after rounding}
}
\value{
a vector of RVs sampled from a unit simplex
}
\description{
This function samples from a simplex using a trick found here - http://blog.geomblog.org/2005/10/sampling-from-simplex.html
}
