% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2.de.R
\name{test2.de}
\alias{test2.de}
\title{Differential expression testing for two cell types}
\usage{
test2.de(y, genotype, measProp2, sd2, priorNormSD = 100, numCellTypes,
  log = T, useHierarchichalModel = F, ...)
}
\arguments{
\item{y}{bulkExpression as a gene by samples matrix}

\item{genotype}{genotype labels for the samples}

\item{measProp2}{estimated proportions of second cell type}

\item{sd2}{standard deviation error of measured proportion}

\item{priorNormSD}{dunno what this is}

\item{log}{log progress}

\item{useHierarchichalModel}{use a hierarchical model instead}

\item{...}{arguments to be passed to `rstan::sampling` (e.g. chains, iter (suggested > 8000), init, verbose, refresh)}

\item{numCelltypes}{number of cell types}

\item{n}{number of samples}
}
\value{
list containing eQTL effect sizes for cancer and normal cell types
}
\description{
This is a refactor of Paul's code.
Limited to two cell types.
NOT to be exported.
}
