% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_bw}
\alias{write_bw}
\title{Write a BigWig track from either a \code{GRanges} or \code{rtracklayer} object}
\usage{
write_bw(x, out, log2_scale = FALSE, unlog2_scale = FALSE)
}
\arguments{
\item{x}{The object to export}

\item{out}{The connection from which data is saved}

\item{log2_scale}{boolean indicating whether to log2 scale intensity values (pseudocount of 1)}

\item{unlog2_scale}{boolean indicating whether to remove log2 scaled intensity values
(assuming pseudocount of 1)}
}
\value{
nothing
}
\description{
Write a BigWig track from either a \code{GRanges} or \code{rtracklayer} object
}
