% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baycon.R
\name{baycon}
\alias{baycon}
\title{Bayesian deconvolution with Stan}
\usage{
baycon(bulkExpression, sigMat, fit.mvn = F, useHyperPrior = F,
  useOptim = F, useADVI = F, stanWarningCheck = "none", ...)
}
\arguments{
\item{bulkExpression}{bulk expression matrix}

\item{sigMat}{signature matrix}

\item{fit.mvn}{a logical indicating whether we want to fit a MVN to the posterior}

\item{useHyperPrior}{use a deconvolution model with hyper prior on the degrees of freedom}

\item{useOptim}{use the L-BFGS optimizer to determine posterior mode}

\item{useADVI}{use the variational bayes approach to estimate posterior mean and variances}

\item{stanWarningCheck}{level at which Stan warnings must be handled, one of \code{c('none', 'divergent', 'all')}}

\item{...}{arguments to be passed to \code{rstan::sampling} (e.g. \code{chains, iter, init, verbose, refresh})}
}
\value{
An object of type list with posterior mean estimates (and MVN mean estimates),
alongwith associated error variances
}
\description{
Bayesian deconvolution with Stan
}
