% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baycon.R
\name{baycon}
\alias{baycon}
\title{Bayesian deconvolution with Stan}
\usage{
baycon(numGenes = nrow(p_bulkExpressionSimMat),
  numCellTypes = ncol(p_simSigMatTwo),
  bulkExpression = p_bulkExpressionSimMat[, 1:5],
  sigMat = p_simSigMatTwo, fit.mvn = F, useHyperPrior = F,
  stanWarningCheck = "none", ...)
}
\arguments{
\item{numGenes}{numeric of number of genes}

\item{numCellTypes}{numeric of number of cell types}

\item{bulkExpression}{bulk expression matrix}

\item{sigMat}{signature matrix}

\item{fit.mvn}{a logical indicating whether we want to fit a MVN to the posterior}

\item{stanWarningCheck}{level at which Stan warnings must be handled, one of \code{c('none', 'divergent', 'all')}}

\item{...}{arguments to be passed to \code{rstan::sampling} (e.g. \code{chains, iter, init, verbose, refresh})}
}
\value{
An object of type vector (list) with posterior mean estimates (and mle mean estimates),
alongwith associated error variances
}
\description{
Bayesian deconvolution with Stan
}
