% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baycon.R
\name{baycon}
\alias{baycon}
\title{Bayesian deconvolution with Stan}
\usage{
baycon(numGenes = nrow(p_bulkExpressionSimMat),
  numCellTypes = ncol(p_simSigMatTwo),
  bulkExpression = p_bulkExpressionSimMat, sigmat = p_simSigMatTwo,
  mle = F, ...)
}
\arguments{
\item{numGenes}{numeric of number of genes}

\item{numCellTypes}{numeric of number of cell types}

\item{bulkExpression}{bulk expression matrix}

\item{sigmat}{signature matrix}

\item{mle}{a logical indicating whether we need mle or not}

\item{...}{arguments to be passed to `rstan::sampling` (e.g. chains, iter, init, verbose, refresh)}
}
\value{
An object of type vector (list) with mean estimates from posterior (and mle estimates)
}
\description{
Bayesian deconvolution with Stan
}
