% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_loci}
\alias{cluster_loci}
\title{Cluster genomic loci via accelerated K-Means clustering}
\usage{
cluster_loci(x, k = 500)
}
\arguments{
\item{x}{a \code{GRanges} object with ChIP-seq signal data with rows as
genomic bins and columns as cell types}

\item{k}{number of k clusters}
}
\value{
a \code{GRanges} object containing genomic coordinates tagged with a
cluster label
}
\description{
Cluster genomic loci via accelerated K-Means clustering
}
