% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSimDataset.R
\name{makeSimDataset}
\alias{makeSimDataset}
\title{Function to make a simple simulated dataset.}
\usage{
makeSimDataset(sdCancer = 1, sdNormal = 1, seed = 12345)
}
\arguments{
\item{sdCancer}{std dev of cancer sample}

\item{sdNormal}{std dev of normal sample}

\item{seed}{set the seed for the simulation}
}
\value{
An object of type list with following names:

\code{cancerExpressionSimMat, normalExpressionSimMat, bulkExpressionSimMat,
simulatedEffectSizesCancer, simulatedEffectSizesNormal}
}
\description{
Function to make a simple simulated dataset.
}
